+++ contracts/protocols/lending/hundred-finance/HfPoolAdapter.sol
@@ -408,8 +408,8 @@ contract HfPoolAdapter is IPoolAdapter, IPoolAdapterInitializerWithAP {
   }
 
   /// @notice Compute current cost of the money
-  function getAPR18() external view override returns (uint) {
-    return IHfCToken(borrowCToken).borrowRatePerBlock() * controller.blocksPerDay() * 365 * 100;
+  function getAPR18() external view override returns (int) {
+    return int(IHfCToken(borrowCToken).borrowRatePerBlock() * controller.blocksPerDay() * 365 * 100);
   }
