+++ contracts/tests/LendingPlatformMock.sol
@@ -69,7 +69,7 @@ contract LendingPlatformMock is IPlatformAdapter {
     return AppDataTypes.ConversionPlan({
       converter: _converter,
       liquidationThreshold18: liquidationThresholds18[collateralAsset_],
-      apr18: borrowRates[borrowAsset_] * countBlocks_,
+      apr18: int(borrowRates[borrowAsset_] * countBlocks_),
       ltv18: liquidationThresholds18[collateralAsset_],
       maxAmountToBorrowBT: liquidity[borrowAsset_],
       maxAmountToSupplyCT: type(uint).max
